Add tests for Range
Let's add some tests for Range extension method. We generate a random
array of floats using Prop.ForAll<float[]>. FsCheck implements default
generators for many, many standard types, so we can use them without an
effort.

A property must be either satisfied or violated. So any boolean is
actually a property. In practice, it is a boolean value returned from a
function that is given required arguments to decide if the property
holds or not.

What are the properties of Range output? Well, maximum value means that
there exists no other value that is greater than maximum, but at the
same time, the maximum must still be a value from the array (not an
arbitrarily high constant). The property for minimum is analogous. Both
these properties require simply the original array and maximum,
respectively minimum, returned from Range.