Consider only non-empty arrays
Our properties fail on empty arrays. That makes sense, because there is
no maximum nor minimum of an empty array. It is reasonable to check the
output only for non-empty arrays.

Thus we add When condition. FsCheck will consider only inputs that
satisfy this condition and ignore the test for the invalid inputs.