Consider only finite values
Now we learned that our Range method does not work when there is a
non-finite value (NaN or infinity). There is not an obvious way how to
handle that, so let's just declare that our Range method works only on
finite values and the output is unspecified otherwise.

So we just add another condition for a valid input generated by FsCheck.
So far, it seems that we are only fixing our test rather than actually
testing our implementation. But that's often the case, during building a
property-based test, you keep discovering violations to your current
properties and you need to decide if that is a bug or a part of the
behavior of your code. In the latter case, the process can even help you
to realize some aspects of your code that might lead to refining your
specification.