Implement custom arbitrary
Argument exhaustion means that FsCheck was not able to generate enough
valid inputs to satisfy number of passing tests configuration before
exhausting the number of allowed attempts. The upper bound for attempts
is a reasonable thing, because if that happens it implies that your
input generation is ill-defined because it generates many invalid
inputs, and you don't want to run forever in that case. It also
indicates that FsCheck quite often generate exceptional values like NaN,
null, crazy unicode and so on, which is generally a reasonable thing to
do in a test framework.

So we need to implement a custom Arbitrary. Arbitrary is a thing that
provides a generator that knows how to generate values of certain type.
You can implement your arbitraries to override the default behavior for
standard types (our case here), but you must implement them for your own
types.

A generator of type T is an instance of Gen<T>. FsCheck provides many
combinators to construct what you want (e.g., Choose to get a random
number from given range, Select to map a value to a new representation,
etc.). It is important to use those instead of standard ways like Random
or LINQ - this will be obvious later. For reference, use the
documentation or autocomplete in your IDE.

In our case, we will generate an array of floats where all values are
finite. The relocation of this condition from test to the generator is
important, because it ensures that only valid values are passed to the
test runner.

There is small inconvenience in using FsCheck in C#. Since the library
is written in F#, often the output of a combinator is an F# type (e.g.,
ListOf returns FSharpList instead of List). So you just need to convert
it into corresponding C# type as you like.