Add tests for Normalize
Ok, now it seems that our Range method works without discovering any
bug. But let's not give up and try to write a test for Normalize method.

An obvious property for Normalize is that all values in the output are
between zero and one. We also mark all inputs that are empty arrays as
trivial instances using Trivial. The effect is that the information of
how much of the inputs were trivial is shown in the output when the test
passes. This is valuable since if a majority of inputs are trivial, it
signals that something is not ok.