Fix Normalize for same-value arrays
Finally a bug! If an array contains only equal values, than there is a
division by zero. So we implement a fix.

Note that if we fixed this failure with only handling one-element
arrays, then FsCheck might really discover that this issue is for
equal-value arrays. But it might also not, since a chance for generating
an array of floats (or even doubles) where all values are equal is
relatively small.