Implement shrinker for custom arbitrary
The actual reason for differentiating between generators and arbitraries
is that arbitraries provide a generator, but also know how to shrink a
generated value to a smaller instance.

Shrinking is extremely useful feature of property-based testing, and is
actually one of key aspects that differentiate it from an ad hoc random
testing (or fuzzing).

When FsCheck discovers a failing input, it tries to reduce it as much as
possible, to provide us with (hopefully) minimal counter-example. Thanks
to that, the counter-example we are given is usually small and
comprehensible enough to being debugged manually.

An FsCheck shrinker is a function that takes a failing input and returns
a collection of instances that are in some sense smaller than the
original value (e.g., number with lower value, array with fwere
elements, etc.).