Demonstrate that there is still a bug
Great, all tests pass. Does it mean that our code is free from bugs?
Unfortunately, no... The problem we encountered in the previous step is
actually more thorough and does not relate only to float's minimum and
maximum only, but to arrays whose range in values is greater than can be
represented in a float. So as it is true for any other approach, test
does only prove the presence of bugs, but not their absence.